ability to turn lights on and off and change their brightness
